reverted to vukkybot master to fix broken crap

+++ commands/say.js
@@ -6,10 +6,13 @@ module.exports = {
 	description: "Make VukkyBot say things!",
 	botPermissions: ["EMBED_LINKS", "MANAGE_MESSAGES"],
 	cooldown: 0,
-	botOwnerOnly: true,
 	execute(message, args) {
-		let say = args.slice(0).join(" ");
-		message.delete();
-		message.channel.send(say);
+		if(!config.misc.owner.includes(message.author.id)) {
+			message.channel.send(embeds.errorEmbed("Sorry, but you're not an owner of this VukkyBot."));
+		} else {
+			let say = args.slice(0).join(" ");
+			message.delete();
+			message.channel.send(say);
+		}
 	},
 };
